/admin/upgrade-versions doesn't reload the page on each update
